requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.26.2
pydantic==2.5.3
pyyaml==6.0.1

# Data visualization
plotly==5.18.0
streamlit>=1.50.0

# Legacy dashboard (optional)
dash==2.14.2
dash-bootstrap-components==1.5.0

# PDF generation
reportlab==4.0.9

# Email
python-dotenv==1.0.0
jinja2==3.1.2
email-validator>=2.0.0

# Logging
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
